+++ UIComponents.Abstractions/Interfaces/IHtmlResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace UIComponents.Abstractions.Interfaces;
+
+public interface IHtmlResponse
+{
+    public string type { get; }
+
+}