+++ UIComponents.Abstractions/Models/HtmlResponse/UICRedirectJson.cs
@@ -1,0 +1,13 @@
+﻿namespace UIComponents.Abstractions.Models.HtmlResponse;
+
+public class UICRedirectJson : IHtmlResponse
+{
+    public string type => "Redirect";
+
+    public UICRedirectJson(string url)
+    {
+        this.url = url;
+    }
+
+    public string url { get; set; }
+}