+++ UIComponents.Generators/Registrations/UICRegistrator.cs
@@ -180,4 +180,5 @@ public static class UICConfigure
         return configOptions;
     }
 
+    
 }